Create EnemySpawner
Enemy Spawner script for Bubble Slime. It uses an array of game objects that I just change based on the global score value. It also has a spawn rate that I change. This could probably be more efficient. Changing the number of objects on the array changes the 1/1 chance of something spawning. So if there are 4 objects and 2 are enemy_one and 2 are enemy_2 then it's basically a 50/50 chance when I run the array through the "Choose" script.